requirements.txt
streamlit
crewai
langchain
langchain-community
langchain-groq
tavily-python
python-dotenv
pydantic
pypdf
sentence-transformers
faiss-cpu
transformers
torch
